+++ packages/remix-cloudflare/CHANGELOG.md
@@ -1,5 +1,12 @@
 # `@remix-run/cloudflare`
 
+## 2.8.1
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/server-runtime@2.8.1`
+
 ## 2.8.0
 
 ### Patch Changes